validate users work

+++ controllers/ProjectController.js
@@ -1,5 +1,6 @@
 import mongoose from 'mongoose';
 import { ProjectModel } from "../models/Project.js";
+import { validateAdmin } from '../JWT.js';
 
 export const getProjects = async (req, res) => { 
     try {
@@ -23,48 +24,63 @@ export const getProject = async (req, res) => {
 export const createProject = async (req, res) => {
     const { projectTitle, author, projectImage, projectLink, projectType, description, repository, projectLead, projectKey } = req.body;
     const newProject = new ProjectModel({ projectTitle, author, projectType, projectImage, projectLink, description, repository, projectLead, projectKey })
-    try {
-        await newProject.save();
-        res.status(201).json("Project Created");
-    } catch (error) {
-        res.status(400).json({ message: error.message });
-    }   
+    let token = req.headers.authorization;
+    if(validateAdmin(token)){
+        try {
+            await newProject.save();
+            res.status(201).json("Project Created!");
+        } catch (error) {
+            res.status(400).json({ message: error.message });
+        }   
+    } else {
+        res.status(400).json('Invalid');
+    }
 }
 
 export const editProject = async (req, res) => {
     const { projectId } = req.params;
     const { projectTitle, startDate, author, projectImage, projectLink, projectType, description, repository, projectLead } = req.body;
-    try {
-        await ProjectModel.findOneAndUpdate(
-            { "_id": projectId },
-            {
-                $set:{
-                    projectTitle: projectTitle,
-                    startDate: startDate,
-                    author: author,
-                    projectImage: projectImage,
-                    projectLink: projectLink,
-                    projectType: projectType,
-                    description: description,
-                    repository: repository,
-                    projectLead: projectLead,
-                }
-            },
-            {new: true}
-        );
-    res.json("Project Updated");
-    } catch (error) {
-        res.status(400).json({ message: error.message });
+    let token = req.headers.authorization;
+    if(validateAdmin(token)){
+        try {
+            await ProjectModel.findOneAndUpdate(
+                { "_id": projectId },
+                {
+                    $set:{
+                        projectTitle: projectTitle,
+                        startDate: startDate,
+                        author: author,
+                        projectImage: projectImage,
+                        projectLink: projectLink,
+                        projectType: projectType,
+                        description: description,
+                        repository: repository,
+                        projectLead: projectLead,
+                    }
+                },
+                {new: true}
+            );
+        res.json("Project Updated");
+        } catch (error) {
+            res.status(400).json({ message: error.message });
+        }
+    } else {
+        res.status(400).json('Invalid');
     }
 }
 
 export const deleteProject = async (req, res) => {
     const { projectId } = req.params;
     if (!mongoose.Types.ObjectId.isValid(projectId)) return res.status(404).send(`No project with id: ${projectId}`);
-    try {
-        await ProjectModel.findByIdAndRemove(projectId);
-        res.json("Project Deleted");
-    } catch (error) {
-        res.status(400).json({ message: error.message });
+    let token = req.headers.authorization;
+    if(validateAdmin(token)){
+        try {
+            await ProjectModel.findByIdAndRemove(projectId);
+            res.json("Project Deleted");
+        } catch (error) {
+            res.status(400).json({ message: error.message });
+        }
+    } else {
+        res.status(400).json('Invalid');
     }
 }
